+++ src/clients/BrowserClient.js
@@ -10,6 +10,8 @@ const Game = require("../utils/Game.js");
 
 const URLFetcher = require("../utils/URLFetcher.js");
 
+const toString = require("../utils/toString.js");
+
 /**
  * The Browser Client Instance for supporting mod codes running in Browser Modding. <br><b>Warning: </b><br><ul><li>This client doesn't support undocumented features like accessing through `game.modding`, etc. </li><li>Some of the latest features of the new ModdingClient (which may not work in browsers) will be available</li>
  * @param {object} options - options for calling the object. <br><b>Note that</b> if both one property and its aliases exist on the object, the value of the main one will be chosen
@@ -34,6 +36,8 @@ class BrowserClient {
 		this.#asynchronous = !!(options?.asynchronous ?? options?.async ?? true);
 		let crashOnError = this.#crashOnError = !!(options?.crashOnException ?? options?.crashOnError);
 		let node = this.#node = new ModdingClient({...options, cacheEvents: true, cacheOptions: false});
+
+		this.#game = new Game(node);
 
 		let handle = function (spec, ...params) {
 			let context = this.#game.modding?.context;
@@ -313,6 +317,42 @@ class BrowserClient {
 		}
 		else if (!this.#node.started) throw new Error("Mod is starting. Please be patience");
 		return await node.start()
+	}
+
+	/**
+	 * Executes any terminal command on current running instance.
+	 * @param {string} command - Command to execute
+	 * @param {boolean} [allowEval = false] - Whether to allow eval the command as JavaScript or not.<br> WARNING: THIS MAY CAUSE SECURITY ISSUE TO YOUR INSTANCE
+	 * @param {boolean} [captureOutput = false] - Whether to capture execution output or pipe it to error/log events instead
+	 * @returns {({ success: boolean, output: any })} Success status (boolean) and output of given execution (if ouput capturing is enabled)
+	 */
+
+	execute (command, allowEval = false, captureOutput = false) {
+		command = toString(command);
+		try {
+			let cmdName = command.trim().split(" ")[0] || "";
+			let cmd, output;
+			if (cmdName && "function" === typeof (cmd = this.#game?.modding?.commands?.[cmdName])) {
+				output = cmd.call(this.#game, command);
+			}
+			else if (!allowEval) {
+				if (!cmdName) throw new Error("No terminal command specified");
+				throw new Error("Unknown terminal command: " + cmdName);
+			}
+			else {
+				output = (function (game, echo, window, global, cmd, cmdName, allowEval, captureOutput) {
+					return eval(command);
+				}).call(this.#game, this.#game, this.#game?.modding?.terminal?.echo, global);
+			}
+
+			if (captureOutput) return { success: true, output };
+			this.#node.log(output);
+			return { success: true };
+		} catch (e) {
+			if (captureOutput) return { success: false, output: e };
+			this.#node.error(e);
+			return { success: false };
+		}
 	}
 
 	/**